backend/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.9,<5.0
psycopg2-binary>=2.9,<3.0
gunicorn>=23.0,<24.0
django-environ>=0.11,<1.0
djangorestframework-simplejwt>=5.4,<6.0
